+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - New #358: Add `ColumnDefinitionBuilder` class (@Tigrov)
 - Enh #359: Refactor `Dsn` class (@Tigrov)
 - Enh #361, #362: Refactor `Schema::findColumns()` method (@Tigrov)
+- Enh #363: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
 
 ## 1.2.0 March 21, 2024